FIR IDE: Move applicators for adding argument names to common location.
